Impacts: * `SampleFactory`: added a reference to new demo `SampleVeryHighZoomLevel` * `TileSystem`: switched to `double` zoom level versions for methods `MapScale` and `GroundResolution` * `TileSystemMathTest`: moved to `TileSystemTest` methods `test_MapSize`, `test_groundResolution` and `test_groundMapScale` * `TileSystemTest`: moved from `TileSystemTest` methods `test_MapSize`, `test_groundResolution` and `test_groundMapScale`; improved their handling of high zoom levels
